refactor(ext): file path change and import update
