+++ extension/api/auth.api.js
@@ -1,4 +1,4 @@
-import { asyncHandler } from "@/lib/asyncHandler.js";
+import { asyncHandler } from "@/entrypoints/utils/asyncHandler.js";
 import { api } from "./base.js";
 
 const fetchAuthUser = async () => {
